refactor: move script payment/spending and asset minting functions to eutxo module
